requirements.txt
backports.cached-property==1.0.1
black==22.1.0
click==8.0.4
graphql-core==3.2.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
Pygments==2.11.2
python-dateutil==2.8.2
python-multipart==0.0.5
sentinel==0.3.0
six==1.16.0
strawberry-graphql==0.99.1
tomli==2.0.1
typing-extensions==4.1.1
